+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### 8.0.6 (08/16/2022 - )
 
+* log-processor: always use bulk index, to simplify
+  > ES already unified bulk index / single index thread pool, there is only one "write" pool
+  > refer to https://www.elastic.co/guide/en/elasticsearch/reference/current/modules-threadpool.html
+
 ### 8.0.5 (07/28/2022 - 08/15/2022)
 
 * mock: fixed MockRedis.list().range(), with negative start or stop